+++ NEWS.md
@@ -12,6 +12,9 @@
 * Packages installed from source within an renv project are not associated
   with repositories. (#1004)
 
+* The configured repositories are searched for packages installed into an renv
+  project from local or unknown sources. (#1004)
+
 # rsconnect 1.1.0
 
 * Fixed analysis of directories that were smaller than the